#54 - IsHeader matcher made public
